# api/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/animals/dewormings/{animalId}": {
      "post": {
        "operationId": "DewormingsController_create",
        "parameters": [
          {
            "name": "animalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDewormingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Dewormings"
        ]
      },
      "get": {
        "operationId": "DewormingsController_list",
        "parameters": [
          {
            "name": "animalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Dewormings"
        ]
      }
    },
    "/animals/{animalId}/dewormings/{dewormingId}": {
      "patch": {
        "operationId": "DewormingsController_update",
        "parameters": [
          {
            "name": "animalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dewormingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDewormingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Dewormings"
        ]
      },
      "delete": {
        "operationId": "DewormingsController_remove",
        "parameters": [
          {
            "name": "animalId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dewormingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Dewormings"
        ]
      }
    },
...
    "/animals": {
      "post": {
        "operationId": "AnimalsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnimalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Animals"
        ]
      }
    },
...
    "/auth/session": {
      "post": {
        "operationId": "AuthController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/google": {
      "post": {
        "operationId": "AuthController_signInWithGoogle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleSignInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/blocks/{id}": {
      "post": {
        "operationId": "BlocksController_block",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Blocks"
        ]
      },
      "delete": {
        "operationId": "BlocksController_unblock",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Blocks"
        ]
      }
    },
...
    "/animals/{id}/adoption-requests": {
      "post": {
        "operationId": "AnimalAdoptionRequestsController_create",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdoptionRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AnimalAdoptionRequests"
        ]
      }
    },
    "/adoption-requests/{id}/approve": {
      "patch": {
        "operationId": "AdoptionRequestsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdoptionRequests"
        ]
      }
    },
    "/adoption-requests/{id}/reject": {
      "patch": {
        "operationId": "AdoptionRequestsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdoptionRequests"
        ]
      }
    },
...
  },
  "info": {
    "title": "PataNet API",
    "description": "API do ecossistema PataNet (usuarios, pets, adocao, eventos, feed e moderacao)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateDewormingDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateDewormingDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateAnimalDto": {
        "type": "object",
        "properties": {}
      },
...
      "SignInDto": {
        "type": "object",
        "properties": {}
      },
      "GoogleSignInDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateAdoptionRequestDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}